Develop

+++ src/main/java/com/projeto/controleanimal/controller/animal/Controller.java
@@ -3,7 +3,10 @@ package com.projeto.controleanimal.controller.animal;
 import com.projeto.controleanimal.model.Animal;
 import com.projeto.controleanimal.model.Cat;
 import com.projeto.controleanimal.service.AnimalService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -13,6 +16,8 @@ import java.util.List;
 public class Controller {
     private final AnimalService service;
 
+    private final String secret = System.getenv("API_SECRET");
+
     public Controller(AnimalService service) {
         this.service = service;
     }
@@ -20,7 +25,7 @@ public class Controller {
 
     @GetMapping("/{id}")
     public Animal getAnimal(@PathVariable("id") Long id) {
-     return service.getAnimal(id);
+        return service.getAnimal(id);
     }
 
     @GetMapping("/all")
@@ -30,7 +35,10 @@ public class Controller {
 
     //TODO checar qual tipo de animal é para criar uma instancia ESPECIFICA
     @PostMapping()
-    public Animal addCat(@RequestBody Animal animal){
+    public Animal addCat(@RequestHeader("x-api-key") String key, @RequestBody Animal animal) {
+        if (!secret.equals(key)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Não autorizado");
+        }
         return service.addAnimal(animal);
     }
 
@@ -38,8 +46,11 @@ public class Controller {
     //ex: http://localhost:8080/api/remove?name=Ivaldo
 
     @DeleteMapping("/remove")
-    public void deleteAnimal(@RequestParam Long id) {
-         service.deleteAnimal(id);
+    public void deleteAnimal(@RequestHeader("x-api-key") String key, @RequestParam Long id) {
+        if (!secret.equals(key)) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Não autorizado");
+        }
+        service.deleteAnimal(id);
     }
 
 }
